zhushou/requirements.txt
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
nltk==3.8.1
spacy==3.7.2
gensim==4.3.2
ipfshttpclient==0.8.0a2
web3==6.11.3
eth-utils==2.3.1
PyPDF2==3.0.1
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
pytest==7.4.3
jupyter==1.0.0
notebook==7.0.6
reportlab==4.0.7

